requirements.txt
bs4==0.0.1
Django==4.2.6
django-cors-headers==4.3.0
django-froala-editor==4.1.3
django-storages @ git+https://github.com/rohanphulkar/django-storages.git@3e7943919e56fe69ad325ee698dc632ae89b5541
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
openai==0.28.1
Pillow==10.1.0
PyJWT==2.8.0
python-decouple==3.8
razorpay==1.4.1
requests==2.31.0
shortuuid==1.0.11
uuid==1.30